requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asari==0.2.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.4.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.115.8
fastjsonschema==2.19.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-DotEnv==0.1.2
Flask-Login==0.6.3
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
googlemaps==4.10.0
greenlet==3.0.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.6
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.1.2
Janome==0.4.2
japanize-matplotlib==1.1.3
jedi==0.19.1
Jinja2==3.1.2
jiter==0.5.0
joblib==1.4.2
Js2Py==0.74
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.1
matplotlib-fontja==1.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mysql-connector-python==9.2.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.26.2
onnx==1.16.2
onnxconverter-common==1.14.0
onnxruntime==1.19.0
openai==1.42.0
openpyxl==3.1.5
optuna==3.6.1
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.6
pillow==10.4.0
pipwin==0.5.2
platformdirs==4.2.2
plotly==5.22.0
prometheus_client==0.20.0
prompt_toolkit==3.0.45
protobuf==3.20.2
psutil==5.9.8
pure-eval==0.2.2
pyarrow==17.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
pyjsparser==2.7.1
PyMySQL==1.1.1
pyocr==0.8.5
pyparsing==3.1.2
PyPrind==2.11.3
pyreadline3==3.4.1
pySmartDL==1.3.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==71.0.4
six==1.16.0
skl2onnx==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.45.3
streamlit==1.37.1
sympy==1.13.2
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.34.0
watchdog==4.0.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
widgetsnbextension==4.0.11
xgboost==2.1.0
yarl==1.9.4
yfinance==0.2.41